setup mysql and migrate first files
Setups to include time zones in mysql and how to migrate your first models.

Follow the "commands to setup myself.txt" files in the root of the project and that should bring you to the same stage as me.

Also, if you want to go through the django tutorial, I'm at the section "Activating models"