created a view and url.  still need to create a view base don the model directly
Hi Robert,

So I don't remember where we left this off for git, but I believe you can simply do a "git clone" into a new folder and it will download everything.

You could delete what you've worked on so far and synchronize with mine in the same folder you are already working on.

I can't seem to find an easy way to simply add the current directory to synchronize with a remote repository so this might be the best way to keep going with this.