Document and Image upload prototype
You can upload documents and images and the link to the database will be made, and a list of all the files is displayed.  We will be able to use this now within an Incident or Scenario to link related files or images for download.  Could also have the image for their profile picture if they want.